+++ overpassforge/__init__.py
@@ -1,3 +1,3 @@
-from .statements import Elements, RawStatement, Nodes, Ways, Relations, Areas, Union, Difference
+from .statements import Elements, RawStatement, Nodes, Ways, Relations, Areas, Union, Difference, OverlappingAreas
 from .builder import Settings, build, beautify
 from .filters import *